This repository hosts a facial recognition model that can train on a set of faces, validate its training, and recognize faces in new images.
Features
•	This model uses Python's pathlib library and the face_recognition library for face detection and recognition tasks.
•	The model can be trained on a set of faces stored in the 'training' directory.
•	The trained model can be validated using a set of images in the 'validation' directory.
•	The model is capable of recognizing faces in new images.
•	Users have the option to choose between 'hog' (CPU) or 'cnn' (GPU) as the model for training and detection tasks.
•	The model uses face_recognition.face_locations() to detect faces in images and face_recognition.face_encodings() to generate encodings for the detected faces. These encodings are numeric representations of facial features, used to match similar faces.
•	The encodings and their corresponding labels are saved to disk using pickle for future recognition tasks.
•	The repository also includes a utility for displaying recognized faces in an image with their labels.

Usage
The primary functions in this repository are:
encode_faces(): This function encodes faces from images in the 'training' directory and saves the encodings and labels to disk.
recognize_faces(): This function recognizes faces in a given image based on the saved encodings and displays the image with the recognized faces labelled.
validate(): This function uses the images in the 'validation' directory to validate the trained model.
The repository can be used as a script with command-line arguments for training, validation, and testing. For example:
python face_recognition_model.py --train --validate --test -f "path_to_image" -m "model_choice"